LayoutTests/editing/async-clipboard/clipboard-read-basic-expected.txt

This test verifies that clipboard items can be read with their expected types, and that the wrapper for clipboard items is the same between calls to Clipboard.read. This test requires WebKitTestRunner or DumpRenderTree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sortedTypes.length is 3
PASS sortedTypes[0] is "text/html"
PASS sortedTypes[1] is "text/plain"
PASS sortedTypes[2] is "text/uri-list"
PASS item.foo is undefined
PASS readCount became 1
PASS sortedTypes.length is 3
PASS sortedTypes[0] is "text/html"
PASS sortedTypes[1] is "text/plain"
PASS sortedTypes[2] is "text/uri-list"
PASS item.foo is "bar"
PASS readCount became 2
PASS successfullyParsed is true

TEST COMPLETE
